use tab layout for better navigation
With current nav setting, it provides no overview of what is covered
in each topic and you are forced to click the "next" link to go through
all the subsections in each topic.

Use tab layout to show all the subsections in the side panel for better
navigation.

also remove the trailing space in each line.