Update vpc-cni.adoc (#660)
I am skeptical of the previous statement made in the documentation where the CNI consumption of IP addresses is said to be 49.

In total, CNI should be taking up 50 IP addresses for nodes + pods.

Now, the previous statement mentioned that it will use 49 IP addresses in warm pool which seemed wrong. (Warm = unused)

That being said, I have added my calculations in the document as well to make it easier.
Primary IPs from each ENI would not be considered under warm pool.
CoreDNS takes up 1 IP each for 2 pods. 
Node 3 doesn't have coreDNS and has one primary IP on primary ENI, so it's left with 9 IPs available for the pods.